# #7527 performance benchmark - verifying performance impact of inferred query parameter types