--- a/tests/Doctrine/Performance/EntityManagerFactory.php
+++ b/tests/Doctrine/Performance/EntityManagerFactory.php
@@ -2,6 +2,10 @@
 
 namespace Doctrine\Performance;
 
+use Doctrine\Common\EventManager;
+use Doctrine\DBAL\Cache\ArrayStatement;
+use Doctrine\DBAL\Cache\QueryCacheProfile;
+use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver\PDOSqlite\Driver;
 use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
@@ -20,7 +24,7 @@ final class EntityManagerFactory
         $config->setAutoGenerateProxyClasses(ProxyFactory::AUTOGENERATE_EVAL);
         $config->setMetadataDriverImpl($config->newDefaultAnnotationDriver([
             realpath(__DIR__ . '/Models/Cache'),
-            realpath(__DIR__ . '/Models/GeoNames')
+            realpath(__DIR__ . '/Models/GeoNames'),
         ], true));
 
         $entityManager = EntityManager::create(
@@ -35,5 +39,30 @@ final class EntityManagerFactory
             ->createSchema(array_map([$entityManager, 'getClassMetadata'], $schemaClassNames));
 
         return $entityManager;
+    }
+
+    public static function makeEntityManagerWithNoResultsConnection() : EntityManagerInterface
+    {
+        $config = new Configuration();
+
+        $config->setProxyDir(__DIR__ . '/../Tests/Proxies');
+        $config->setProxyNamespace('Doctrine\Tests\Proxies');
+        $config->setAutoGenerateProxyClasses(ProxyFactory::AUTOGENERATE_EVAL);
+        $config->setMetadataDriverImpl($config->newDefaultAnnotationDriver([
+            realpath(__DIR__ . '/Models/Cache'),
+            realpath(__DIR__ . '/Models/Generic'),
+            realpath(__DIR__ . '/Models/GeoNames'),
+        ], true));
+
+        // A connection that doesn't really do anything
+        $connection = new class ([], new Driver(), null, new EventManager()) extends Connection
+        {
+            public function executeQuery($query, array $params = [], $types = [], QueryCacheProfile $qcp = null)
+            {
+                return new ArrayStatement([]);
+            }
+        };
+
+        return EntityManager::create($connection, $config);
     }
 }